# art_admin_backend/docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateMoodRecordRequest:
    properties:
      activities:
        items:
          type: string
        type: array
      intensity:
        maximum: 10
        minimum: 1
        type: integer
      location:
        type: string
      mood_type:
        type: string
      note:
        type: string
      triggers:
        items:
          type: string
        type: array
    required:
    - intensity
    - mood_type
    type: object
...
  request.MoodAnalyticsRequest:
    properties:
      analyze_type:
        enum:
        - daily
        - weekly
        - monthly
        type: string
      end_date:
        type: string
      start_date:
        type: string
    required:
    - analyze_type
    - end_date
    - start_date
    type: object
...
  request.UpdateMoodRecordRequest:
    properties:
      activities:
        items:
          type: string
        type: array
      id:
        type: integer
      intensity:
        maximum: 10
        minimum: 1
        type: integer
      location:
        type: string
      mood_type:
        type: string
      note:
        type: string
      triggers:
        items:
          type: string
        type: array
    required:
    - id
    type: object
...
  response.MoodAnalyticsItem:
    properties:
      analyze_date:
        type: string
      analyze_type:
        type: string
      average_intensity:
        type: number
      created_at:
        type: string
      id:
        type: integer
      mood_distribution:
        additionalProperties:
          type: integer
        type: object
      most_common_mood:
        type: string
      most_common_mood_label:
        type: string
      total_records:
        type: integer
      user_id:
        type: integer
    type: object
  response.MoodRecordItem:
    properties:
      activities:
        items:
          type: string
        type: array
      created_at:
        type: string
      id:
        type: integer
      intensity:
        type: integer
      location:
        type: string
      mood_label:
        type: string
      mood_type:
        type: string
      note:
        type: string
      triggers:
        items:
          type: string
        type: array
      user_id:
        type: integer
    type: object
  response.MoodStatistics:
    properties:
      average_intensity:
        type: number
      mood_distribution:
        additionalProperties:
          type: integer
        type: object
      most_common_mood:
        type: string
      most_common_mood_label:
        type: string
      recent_trend:
        description: 最近7天趋势
        items:
          $ref: '#/definitions/response.MoodRecordItem'
        type: array
      total_records:
        type: integer
      weekly_comparison:
        additionalProperties: true
        description: 周对比
        type: object
    type: object
...
  response.PaginatedData:
    properties:
      current:
        description: 当前页码
        type: integer
      records:
        description: 记录列表
      size:
        description: 每页条数
        type: integer
      total:
        description: 总记录数
        type: integer
    type: object
  response.Response:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 数据
      msg:
        description: 消息
        type: string
    type: object
...
host: localhost:48080
info:
  contact: {}
  description: Art Admin Backend API Documentation
  title: Art Admin API
  version: "1.0"
paths:
...
  /api/mood-record/analytics:
    post:
      consumes:
      - application/json
      description: 获取用户情绪深度分析
      parameters:
      - description: 分析请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.MoodAnalyticsRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 分析成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MoodAnalyticsItem'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取情绪分析
      tags:
      - 情绪管理
  /api/mood-record/create:
    post:
      consumes:
      - application/json
      description: 创建新的情绪记录
      parameters:
      - description: 情绪记录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateMoodRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 创建情绪记录
      tags:
      - 情绪管理
  /api/mood-record/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 删除情绪记录
      parameters:
      - description: 记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 删除情绪记录
      tags:
      - 情绪管理
  /api/mood-record/list:
    get:
      consumes:
      - application/json
      description: 分页查询情绪记录列表
      parameters:
      - description: 当前页码
        in: query
        minimum: 1
        name: current
        required: true
        type: integer
      - description: 每页条数
        in: query
        maximum: 100
        minimum: 1
        name: size
        required: true
        type: integer
      - description: 用户ID
        in: query
        name: user_id
        type: integer
      - description: 情绪类型
        in: query
        name: mood_type
        type: string
      - description: 开始日期
        in: query
        name: start_date
        type: string
      - description: 结束日期
        in: query
        name: end_date
        type: string
      - description: 最小强度
        in: query
        maximum: 10
        minimum: 1
        name: min_intensity
        type: integer
      - description: 最大强度
        in: query
        maximum: 10
        minimum: 1
        name: max_intensity
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PaginatedData'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取情绪记录列表
      tags:
      - 情绪管理
  /api/mood-record/statistics:
    get:
      consumes:
      - application/json
      description: 获取用户情绪统计信息
      parameters:
      - description: 用户ID
        in: query
        name: user_id
        required: true
        type: integer
      - default: 30
        description: 统计天数
        in: query
        name: days
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 查询成功
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.MoodStatistics'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 获取情绪统计
      tags:
      - 情绪管理
  /api/mood-record/update:
    put:
      consumes:
      - application/json
      description: 更新情绪记录信息
      parameters:
      - description: 情绪记录信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateMoodRecordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: 更新情绪记录
      tags:
      - 情绪管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"